fix: filling data in promo.csv, refactor
